--- src/main/java/com/myFitness/MyFitnessProgramApplication.java
@@ -1,13 +1,0 @@
-package com.myfitness;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MyFitnessProgramApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MyFitnessProgramApplication.class, args);
-	}
-
-}
