--- src/main/java/com/myFitness/persistence/BoardRepository.java
@@ -1,9 +1,0 @@
-package com.myFitness.persistence;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.myFitness.domain.Board;
-
-public interface BoardRepository extends JpaRepository<Board, Long> {
-
-}
