--- src/main/java/com/myFitness/persistence/ClassDiaryRepository.java
@@ -1,9 +1,0 @@
-package com.myFitness.persistence;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.myFitness.domain.ClassDiary;
-
-public interface ClassDiaryRepository extends CrudRepository<ClassDiary, Integer> {
-
-}
