--- src/main/java/com/myFitness/persistence/ReservationRepository.java
@@ -1,9 +1,0 @@
-package com.myFitness.persistence;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.myFitness.domain.Reservation;
-
-public interface ReservationRepository extends CrudRepository<Reservation, Integer> {
-
-}
