--- src/main/java/com/myFitness/service/BoardService.java
@@ -1,19 +1,0 @@
-package com.myFitness.service;
-
-import java.util.List;
-
-import com.myFitness.domain.Board;
-
-public interface BoardService {
-
-	List<Board> getBoardList(Board board);
-
-	void insertBoard(Board board);
-
-	Board getBoard(Board board);
-
-	void updateBoard(Board board);
-
-	void deleteBoard(Board board);
-
-}
