+++ src/main/java/com/myfitness/MyFitnessProgramApplication.java
@@ -1,0 +1,12 @@
+package com.myfitness;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MyFitnessProgramApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MyFitnessProgramApplication.class, args);
+	}
+}
