+++ src/main/java/com/myfitness/domain/Role.java
@@ -1,0 +1,12 @@
+package com.myfitness.domain;
+
+/*
+ *  Security 권한 지정을 위한 Enum 클래스 
+ */
+
+public enum Role {
+	ROLE_MEMBER,    // 일반회원
+	ROLE_PTMEMBER,  // PT등록 회원
+	ROLE_TRAINER,   // 트레이너
+	ROLE_ADMIN      // 관리자
+}
