+++ src/main/java/com/myfitness/persistence/ClassDiaryRepository.java
@@ -1,0 +1,9 @@
+package com.myfitness.persistence;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.myfitness.domain.ClassDiary;
+
+public interface ClassDiaryRepository extends CrudRepository<ClassDiary, Integer> {
+
+}
