+++ src/main/java/com/myfitness/persistence/DietDiaryRepository.java
@@ -1,0 +1,10 @@
+package com.myfitness.persistence;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.myfitness.domain.DietDiary;
+
+
+public interface DietDiaryRepository extends CrudRepository<DietDiary, Long>{
+
+}
