+++ src/main/java/com/myfitness/persistence/ReservationRepository.java
@@ -1,0 +1,9 @@
+package com.myfitness.persistence;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.myfitness.domain.Reservation;
+
+public interface ReservationRepository extends CrudRepository<Reservation, Integer> {
+
+}
