+++ src/main/java/com/myfitness/service/BoardService.java
@@ -1,0 +1,19 @@
+package com.myfitness.service;
+
+import java.util.List;
+
+import com.myfitness.domain.Board;
+
+public interface BoardService {
+
+	List<Board> getBoardList(Board board);
+
+	void insertBoard(Board board);
+
+	Board getBoard(Board board);
+
+	void updateBoard(Board board);
+
+	void deleteBoard(Board board);
+
+}
