--- src/test/java/com/myFitness/MyFitnessProgramApplicationTests.java
@@ -1,13 +1,0 @@
-package com.myFitness;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MyFitnessProgramApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
