+++ src/test/java/com/myfitness/MyFitnessProgramApplicationTests.java
@@ -1,0 +1,13 @@
+package com.myfitness;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MyFitnessProgramApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
